avoid adding spi dependency to common
